Rename v1kubeflow package from sparkManger as unneeded
