+++ internal/sparkManager/api/router.go
@@ -5,7 +5,7 @@ import (
 	"github.com/slackhq/spark-gateway/internal/shared/config"
 	sgMiddleware "github.com/slackhq/spark-gateway/internal/shared/middleware"
 	"github.com/slackhq/spark-gateway/internal/sparkManager/api/health"
-	v1kubeflow "github.com/slackhq/spark-gateway/internal/sparkManager/api/v1/kubeflow"
+	"github.com/slackhq/spark-gateway/internal/sparkManager/api/v1"
 	"github.com/slackhq/spark-gateway/internal/sparkManager/service"
 )
 
@@ -22,7 +22,7 @@ func NewRouter(sgConf *config.SparkGatewayConfig, appService service.SparkApplic
 	// Versioned routes
 	v1Group := router.Group("/api/v1")
 
-	v1kubeflow.RegisterKubeflowApplicationRoutes(v1Group, sgConf, appService)
+	v1.RegisterKubeflowApplicationRoutes(v1Group, sgConf, appService)
 
 	return router, nil
