+++ internal/sparkManager/api/v1/application_handler_test.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package v1kubeflow
+package v1
 
 import (
 	"bytes"
